// compiler/types.ts
export enum TokenType {
  IDENTIFICATION, DIVISION, PROGRAM_ID,
  ENVIRONMENT, CONFIGURATION, INPUT_OUTPUT, FILE_CONTROL, SELECT, ASSIGN, ORGANIZATION,
  LINE, SEQUENTIAL, INDEXED, ACCESS, MODE, DYNAMIC, RANDOM, RECORD, KEY, STATUS,
  DATA, FILE, FD, WORKING_STORAGE, LINKAGE, SECTION, PROCEDURE,
  PIC, VALUE,
  MOVE, ADD, SUBTRACT, MULTIPLY, DIVIDE, 
  TO, FROM, BY, INTO, GIVING, COMPUTE,
  IF, THEN, ELSE, END_IF,
  PERFORM, UNTIL, TIMES, END_PERFORM,
  ACCEPT, DISPLAY, STOP, RUN, CALL, USING, EXIT, PROGRAM, GOBACK,
  OPEN, CLOSE, READ, WRITE, REWRITE, DELETE, START, 
  INPUT, OUTPUT, I_O, EXTEND, AT, END, INVALID, NOT,
  IS,
  // CICS & BMS
  EXEC, CICS, END_EXEC,
  SEND, RECEIVE, MAP, MAPSET,
  TRANSID, COMMAREA, LENGTH, RIDFLD,
  HANDLE, CONDITION,
  LINK, RETURN_CICS, // RETURN is keyword, using RETURN_CICS for internal differentiation if needed, but parser uses context
  MAP_SECTION, DFHMDF, POS, NAME, // BMS keywords
  
  IDENTIFIER, LITERAL_STRING, LITERAL_NUMBER,
  DOT, EQUALS, GREATER, LESS,
  PLUS, MINUS, ASTERISK, SLASH, POWER,
  LPAREN, RPAREN, COLON, // Added Colon for slicing
  ZERO, SPACE,
  EOF
}

export type Severity = 'INFO' | 'WARNING' | 'ERROR' | 'SEVERE';

export interface Token {
  type: TokenType;
  value: string;
  line: number;
  column: number;
}

export interface Diagnostic {
  line: number;
  column: number;
  code: string;
  message: string;
  severity: Severity;
  source?: string;
}